# fix(telnet-host): properly crash; crash fails tests